refactor */pickers.py
