+++ src/faebryk/libs/picker/jlcpcb/pickers.py
@@ -3,6 +3,7 @@ import logging
 import faebryk.library._F as F
 import faebryk.libs.picker.jlcpcb.picker_lib as P
 from faebryk.core.module import Module
+from faebryk.libs.picker.common import StaticPartPicker
 from faebryk.libs.picker.jlcpcb.jlcpcb import JLCPCB_DB, ComponentQuery
 from faebryk.libs.picker.picker import PickError
 
@@ -19,37 +20,13 @@ class JLCPCBPicker(F.has_multi_picker.FunctionPicker):
             raise PickError(e.args[0], module) from e
 
 
-class StaticJLCPCBPartPicker(F.has_multi_picker.Picker):
+class StaticJLCPCBPartPicker(StaticPartPicker):
     """
     Use this if you want to specify a specific part to the multi-picker
     eg. a default switch
     """
 
-    def __init__(
-        self,
-        *,
-        mfr: str | None = None,
-        mfr_pn: str | None = None,
-        lcsc_pn: str | None = None,
-    ) -> None:
-        super().__init__()
-        self.mfr = mfr
-        self.mfr_pn = mfr_pn
-        self.lcsc_pn = lcsc_pn
-
-    def _friendly_description(self) -> str:
-        desc = ""
-        if self.mfr:
-            desc += f"mfr={self.mfr}"
-        if self.mfr_pn:
-            desc += f"mfr_pn={self.mfr_pn}"
-        if self.lcsc_pn:
-            desc += f"lcsc_pn={self.lcsc_pn}"
-        if not desc:
-            return "<no params>"
-        return desc
-
-    def pick(self, module: Module):
+    def _find_parts(self, module: Module):
         q = ComponentQuery()
 
         if self.mfr:
@@ -59,25 +36,7 @@ class StaticJLCPCBPartPicker(F.has_multi_picker.Picker):
         if self.lcsc_pn:
             q.filter_by_lcsc_pn(self.lcsc_pn)
 
-        parts = q.get()
-
-        if len(parts) > 1:
-            raise PickError(
-                f"Multiple parts found for {self._friendly_description()}", module
-            )
-
-        if len(parts) < 1:
-            raise PickError(
-                f"Could not find part for {self._friendly_description()}", module
-            )
-
-        try:
-            parts[0].attach(module, [])
-            return
-        except ValueError as e:
-            raise PickError(
-                f"Could not attach part for {self._friendly_description()}", module
-            ) from e
+        return q.get()
 
 
 def add_jlcpcb_pickers(module: Module, base_prio: int = 0) -> None:
